Add support to the JIT for true non-lazy operation.  When a call to a function that has not been JIT'd yet, the callee is put on a list of pending functions to JIT.  The call is directed through a stub, which is updated with the address of the function after it has been JIT'd.  A new interface for allocating and updating empty stubs is provided.
Add support for removing the ModuleProvider the JIT was created with, which
would otherwise invalidate the JIT's PassManager, which is initialized with the
ModuleProvider's Module.

Add support under a new ExecutionEngine flag for emitting the infomration 
necessary to update Function and GlobalVariable stubs after JITing them, by
recording the address of the stub and the name of the GlobalValue.  This allows
code to be copied from one address space to another, where libraries may live
at different virtual addresses, and have the stubs updated with their new
correct target addresses.

llvm-svn=64906