instcombine transforms vector loads that are only used by extractelement operations into a bitcast of the pointer, then a gep, then a scalar load.  Disable this when the vector only has one element, because it leads to infinite loops in instcombine (PR4908).
This transformation seems like a really bad idea to me, as it
will likely disable CSE of vector load/stores etc and can be
better done in the code generator when profitable.  This
goes all the way back to the first days of packed types,
r25299 specifically.

I'll let those people who care about the performance of vector
code decide what to do with this.

llvm-svn=81185