Make dragonegg capable of building clang.  The problem was the following: a GCC record type can occur in several variants.  When converting records to LLVM types we convert the main variant.  During the type conversion, the index of the LLVM field corresponding to each GCC field is cached in the GCC field.  However, these are the fields of the main variant, not of the original type.  In fact the fields of type variants are never the same as the fields of the main variant, because their DECL_CONTEXT points to the containing type, which differs by definition between variants.  As a result, if we had a COMPONENT_REF with a field of a variant, there would be no LLVM index cached in the field, and we would fall over with an assertion failure. For example, when compiling clang.  Curiously llvm-gcc does not fail on this, yet it has exactly the same bug.  Presumably this means that gcc-4.2 never does a COMPONENT_REF on a type variant.  Indeed gcc-4.5 hardly ever does it either - it takes weird template code to produce an example (the testcase is CodeCompleteConsumer.ii in the dragonegg testsuite).
There are several ways of fixing this.  I chose a rather exotic way: don't
have the type conversion logic calculate LLVM indices for fields at all.
Instead, compute them (and cache them) on the fly when needed.  The result
should be exactly the same as before (less the crashes), except for bitfields:
if a bitfield does not start in the first byte of the LLVM field then you get
worse code.  I plan to do something about this later.

llvm-svn=97803