Generalize tblgen's dag parsing logic to handle arbitrary expressions as the operator of the dag.  Specifically, this allows parsing things like (F.x 4) in addition to just (a 4).
Unfortunately, this runs afoul of an idiom being used by llvmc.  It
is using dags like (foo [1,2,3]) to represent a list of stuff being
passed into foo.  With this change, this is parsed as a [1,2,3] 
subscript on foo instead of being the first argument to the dag.
Cope with this in the short term by requiring a "-llvmc-temp-hack"
argument to tblgen to get the old parsing behavior.

llvm-svn=115742