Clean up the helpers used to compute the usual arithmetic conversions' result type. Localize all of the logic within a single function rather than spreading it throughout the class.
Also fixes a buglet where we failed to check for a RHS arithmetic type wider
than the LHS and return its canonical type. I've yet to produce a test case
that breaks because of this, but it was spotted by inspection by folks on the
IRC channel and is obviously correct now.

llvm-svn=121633