Undid most changes to ExecutionContext as they affected ExecutionContext objects that were being used when the process was privately stopped.
Modified the ExecutionContextRef::Lock() to take a boolean value that can say if you want no thread/frame if the process is running and we now use this in the command interpreter.

Added Point, Size and Rect structs into the curses namespace for much easier API usage. Also added functions that know how to split bounding rectangles horizontally and vertically.

Changed the curses window creation over to use the new Rect functionality.

llvm-svn=198517