Merging rr310989: ------------------------------------------------------------------------ r310989 | mstorsjo | 2017-08-15 22:13:25 -0700 (Tue, 15 Aug 2017) | 18 lines
[COFF] Fix the name type for stdcall functions in import libraries

Since SVN r303491 and r304573, LLD used the COFFImportLibrary
functions from LLVM. These only had two names, Name and ExtName,
which wasn't enough to convey all the details of stdcall functions.

Stdcall functions got the wrong symbol name in the import library
itself in r303491, which is why it was reverted in r304561. When
re-landed and fixed in r304573 (after adding a test in r304572),
the symbol name itself in the import library ended up right, but the
name type of the import library entry was wrong.

This had the effect that linking to the import library succeeded
(contrary to in r303491, where linking to such an import library
failed), but at runtime, the symbol wouldn't be found in the DLL
(since the caller linked to the stdcall decorated name).

Differential Revision: https://reviews.llvm.org/D36545
------------------------------------------------------------------------

llvm-svn=311098