Merging r323395, r323396, r323399, r323440, r323449, r323456, and r323625:
------------------------------------------------------------------------
r323395 | rafael | 2018-01-25 02:29:15 +0100 (Thu, 25 Jan 2018) | 1 line

Use lookup instead of find. NFC, just simpler.
------------------------------------------------------------------------

------------------------------------------------------------------------
r323396 | rafael | 2018-01-25 02:36:36 +0100 (Thu, 25 Jan 2018) | 3 lines

Only lookup LMARegion once. NFC.

This is similar to how we handle MemRegion.
------------------------------------------------------------------------

------------------------------------------------------------------------
r323399 | rafael | 2018-01-25 03:18:00 +0100 (Thu, 25 Jan 2018) | 3 lines

Remove MemRegionOffset. NFC.

We can just use a member variable in MemoryRegion.
------------------------------------------------------------------------

------------------------------------------------------------------------
r323440 | rafael | 2018-01-25 17:43:49 +0100 (Thu, 25 Jan 2018) | 1 line

Simplify. NFC.
------------------------------------------------------------------------

------------------------------------------------------------------------
r323449 | rafael | 2018-01-25 18:42:03 +0100 (Thu, 25 Jan 2018) | 9 lines

Improve LMARegion handling.

This fixes the crash reported at PR36083.

The issue is that we were trying to put all the sections in the same
PT_LOAD and crashing trying to write past the end of the file.

This also adds accounting for used space in LMARegion, without it all
3 PT_LOADs would have the same physical address.
------------------------------------------------------------------------

------------------------------------------------------------------------
r323456 | rafael | 2018-01-25 20:02:08 +0100 (Thu, 25 Jan 2018) | 19 lines

Move LMAOffset from the OutputSection to the PhdrEntry. NFC.

If two sections are in the same PT_LOAD, their relatives offsets,
virtual address and physical addresses are all the same.

I initially wanted to have a single global LMAOffset, on the
assumption that every ELF file was in practiced loaded contiguously in
both physical and virtual memory.

Unfortunately that is not the case. The linux kernel has:

  LOAD           0x200000 0xffffffff81000000 0x0000000001000000 0xced000 0xced000 R E 0x200000
  LOAD           0x1000000 0xffffffff81e00000 0x0000000001e00000 0x15f000 0x15f000 RW  0x200000
  LOAD           0x1200000 0x0000000000000000 0x0000000001f5f000 0x01b198 0x01b198 RW  0x200000
  LOAD           0x137b000 0xffffffff81f7b000 0x0000000001f7b000 0x116000 0x1ec000 RWE 0x200000

The delta for all but the third PT_LOAD is the same:
0xffffffff80000000. I think the 3rd one is a hack for implementing per
cpu data, but we can't break that.
------------------------------------------------------------------------

------------------------------------------------------------------------
r323625 | rafael | 2018-01-29 04:44:44 +0100 (Mon, 29 Jan 2018) | 10 lines

Put the header in the first PT_LOAD even if that PT_LOAD has a LMAExpr.

This should fix PR36017.

The root problem is that we were creating a PT_LOAD just for the
header. That was technically valid, but inconvenient: we should not be
making the ELF discontinuous.

The solution is to allow a section with LMAExpr to be added to a
PT_LOAD if that PT_LOAD doesn't already have a LMAExpr.
------------------------------------------------------------------------

llvm-svn=323733